family_tree_backend/requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==3.2.21
django-cors-headers==4.2.0
django-environ==0.11.2
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.3.0
gunicorn==21.2.0
numpy==1.26.2
opencv-python==4.8.1.78
packaging==23.1
psycopg2-binary==2.9.7
pydocstyle==6.3.0
PyJWT==1.7.1
python-decouple==3.8
pytz==2023.3.post1
snowballstemmer==2.2.0
sqlparse==0.4.4
typing_extensions==4.7.1
